Creating an object sets the viewingDirection
