+++ app/services/cocina/mapper.rb
@@ -6,15 +6,18 @@ module Cocina
   class Mapper
     # Raised when called on something other than an item or collection
     class UnsupportedObjectType < StandardError; end
-
+    # @param [Dor::Abstract] item the Fedora object to convert to a cocina object
+    # @return [Cocina::Models::DRO,Cocina::Models::Collection,Cocina::Models::AdminPolicy]
     def self.build(item)
       new(item).build
     end
 
+    # @param [Dor::Abstract] item the Fedora object to convert to a cocina object
     def initialize(item)
       @item = item
     end
 
+    # @return [Cocina::Models::DRO,Cocina::Models::Collection,Cocina::Models::AdminPolicy]
     def build
       klass = cocina_klass
       props = if klass == Cocina::Models::DRO
@@ -38,15 +41,10 @@ module Cocina
         version: item.current_version,
         administrative: build_administrative,
         identification: build_identification,
-        access: build_access
+        access: AccessBuilder.build(item),
+        structural: DroStructuralBuilder.build(item)
       }.tap do |props|
         props[:identification][:catalogLinks] = [{ catalog: 'symphony', catalogRecordId: item.catkey }] if item.catkey
-
-        unless item.is_a?(Dor::Etd) || item.contentMetadata.new?
-          props[:structural] = {
-            contains: build_filesets(item.contentMetadata, version: item.current_version, id: item.pid)
-          }
-        end
 
         description = build_descriptive
         props[:description] = description unless description.nil?
@@ -62,7 +60,7 @@ module Cocina
         administrative: build_administrative,
         description: build_descriptive,
         identification: build_identification,
-        access: build_access
+        access: AccessBuilder.build(item)
       }.tap do |props|
         props[:identification][:catalogLinks] = [{ catalog: 'symphony', catalogRecordId: item.catkey }] if item.catkey
         description = build_descriptive
@@ -89,7 +87,7 @@ module Cocina
 
     def dro_type
       case item.content_type_tag
-      when 'Image'
+      when 'Image (ltr)', 'Image (rtl)'
         Cocina::Models::Vocab.image
       when '3D'
         Cocina::Models::Vocab.three_dimensional
@@ -104,13 +102,6 @@ module Cocina
       else
         Cocina::Models::Vocab.object
       end
-    end
-
-    # Map values from dor-services
-    # https://github.com/sul-dlss/dor-services/blob/b9b4768eac560ef99b4a8d03475ea31fe4ae2367/lib/dor/datastreams/rights_metadata_ds.rb#L221-L228
-    # to https://github.com/sul-dlss/cocina-models/blob/master/docs/maps/DRO.json#L102
-    def access_rights
-      item.rights.sub('None', 'citation-only').downcase
     end
 
     def build_identification
@@ -129,32 +120,6 @@ module Cocina
         { title: [{ primary: true, titleFull: item.properties.title.first }] }
       else
         { title: [{ primary: true, titleFull: item.full_title }] }
-      end
-    end
-
-    def build_filesets(content_metadata_ds, version:, id:)
-      content_metadata_ds.ng_xml.xpath('//resource').map do |resource_node|
-        files = build_files(resource_node.xpath('file'), version: version, parent_id: id)
-        structural = {}
-        structural[:contains] = files if files.present?
-        Cocina::Models::FileSet.new(
-          externalIdentifier: resource_node['id'],
-          type: Cocina::Models::Vocab.fileset,
-          label: resource_node.xpath('label').text,
-          version: version,
-          structural: structural
-        )
-      end
-    end
-
-    def build_files(file_nodes, version:, parent_id:)
-      file_nodes.map do |node|
-        Cocina::Models::File.new(
-          externalIdentifier: "#{parent_id}/#{node['id']}",
-          type: Cocina::Models::Vocab.file,
-          label: node['id'],
-          version: version
-        )
       end
     end
 
@@ -184,31 +149,6 @@ module Cocina
           release: node.text
         }
       end
-    end
-
-    def build_access
-      { access: access_rights }.tap do |access|
-        embargo = build_embargo
-        access[:embargo] = embargo unless embargo.empty?
-      end
-    end
-
-    def build_embargo
-      return {} unless item.respond_to?(:embargoMetadata) && item.embargoMetadata.release_date
-
-      {
-        releaseDate: item.embargoMetadata.release_date.iso8601,
-        access: build_embargo_access
-      }
-    end
-
-    def build_embargo_access
-      access_node = item.embargoMetadata.release_access_node.xpath('//access[@type="read"]/machine/*[1]').first
-      return 'dark' if access_node.nil?
-      return 'world' if access_node.name == 'world'
-      return access_node.content if access_node.name == 'group'
-
-      'dark'
     end
 
     # @todo This should have more speicific type such as found in identityMetadata.objectType
